chatbot_api/requirements.txt
fastapi==0.121.0
pydantic==2.12.3
uvicorn==0.38.0
langchain==1.0.3
langchain-core==1.0.2
langchain-ollama==1.0.0
langchain-chroma==1.0.0
rich==14.2.0
spacy==3.8.7
chromadb==1.3.0
python-dotenv==1.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl